fix: unnecessary things

+++ src/integrations/ring/client/mod.rs
@@ -34,9 +34,7 @@ pub fn camera_recordings_list(recordings: VideoSearchRes) -> String {
     "<ul>"
         .chars()
         .chain(recordings.video_search.iter().flat_map(|recording| {
-            let created_at_utc = Utc
-                .timestamp_millis_opt(recording.created_at as i64)
-                .unwrap();
+            let created_at_utc = Utc.timestamp_millis_opt(recording.created_at).unwrap();
             let created_at_eastern = created_at_utc.with_timezone(&Eastern);
             let formatted_time = created_at_eastern.format("%B %e, %Y, %I:%M %p").to_string();
